requirements.txt
Django==4.1.2
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
psycopg2-binary
django_redis
django-environ
cryptography
django-rest-knox
django-extensions
requests
user-agents
Pillow
django_filter
daphne==4.0.0
channels==4.0.0
channels-redis
asgi_redis